+++ examples/random.rs
@@ -20,21 +20,17 @@ fn setup(
     mut meshes: ResMut<Assets<Mesh>>,
     mut materials: ResMut<Assets<StandardMaterial>>,
 ) {
-    let mut camera = Camera3dBundle {
-        tonemapping: Tonemapping::None,
-        ..default()
-    };
-    camera.transform.translation = Vec3::new(0.0, 0.0, 100.0);
-    commands.spawn(camera);
+    commands.spawn((
+        Transform::from_translation(Vec3::Z * 100.),
+        Camera3d::default(),
+        Tonemapping::None,
+    ));
 
-    commands.spawn(DirectionalLightBundle {
-        directional_light: DirectionalLight {
-            color: Color::WHITE,
-            // Crank the illuminance way (too) high to make the reference cube clearly visible
-            illuminance: 100000.,
-            shadows_enabled: false,
-            ..Default::default()
-        },
+    commands.spawn(DirectionalLight {
+        color: Color::WHITE,
+        // Crank the illuminance way (too) high to make the reference cube clearly visible
+        illuminance: 100000.,
+        shadows_enabled: false,
         ..Default::default()
     });
 
@@ -95,10 +91,6 @@ fn setup(
         ))
         .with_children(|p| {
             // Reference cube to visualize the emit origin
-            p.spawn(PbrBundle {
-                mesh: cube.clone(),
-                material: mat.clone(),
-                ..Default::default()
-            });
+            p.spawn((Mesh3d(cube.clone()), MeshMaterial3d(mat.clone())));
         });
 }