+++ src/asset.rs
@@ -1,7 +1,7 @@
 use std::ops::Deref;
 
 #[cfg(feature = "serde")]
-use bevy::asset::{io::Reader, AssetLoader, AsyncReadExt, LoadContext};
+use bevy::asset::{io::Reader, AssetLoader, LoadContext};
 use bevy::{
     asset::{Asset, Handle},
     prelude::Mesh,
@@ -1006,11 +1006,11 @@ impl AssetLoader for EffectAssetLoader {
 
     type Error = EffectAssetLoaderError;
 
-    async fn load<'a>(
-        &'a self,
-        reader: &'a mut Reader<'_>,
-        _settings: &'a Self::Settings,
-        _load_context: &'a mut LoadContext<'_>,
+    async fn load(
+        &self,
+        reader: &mut dyn Reader,
+        _settings: &Self::Settings,
+        _load_context: &mut LoadContext<'_>,
     ) -> Result<Self::Asset, Self::Error> {
         let mut bytes = Vec::new();
         reader.read_to_end(&mut bytes).await?;